*: refactor  mount-specific changes while creating a Sandbox container
